cast to float

+++ rt-drone/controls/src/quadrotor_controller.cpp
@@ -16,14 +16,14 @@ void quadRotorController::control_to_state(struct state_struct current_state, st
 	thrust = thrustPID.calculatePID(desired_state.thrust - current_state.thrust);
 
 	printf("controls logging to rpy file. \n");
-	fprintf(rpy_debug, "%f %f %f %f \n", roll, pitch, yaw, thrust);
+	fprintf(rpy_debug, "%f %f %f %f \n", (float)roll, (float)pitch, (float)yaw, (float)thrust);
 
 	//Need to convert then into vel space because that is what the motor speed controls. Maybe. Try this first.
 	
 	// Transform to motor space here
 	Eigen::Vector4f motor_inputs = states_to_motors_transform(roll, pitch, yaw, thrust);
 	printf("controls logging to motor debug. \n");
-	fprintf(motor_debug, "%f %f %f %f\n", motor_inputs(0), motor_inputs(1), motor_inputs(2), motor_inputs(3));
+	fprintf(motor_debug, "%f %f %f %f\n", (float)motor_inputs(0), (float)motor_inputs(1), (float)motor_inputs(2), (float)motor_inputs(3));
 
 	// Output is duty cycle vector float
